requirements.txt
facenet_pytorch==2.5.3
matplotlib==3.10.7
numpy==2.3.5
Pillow==12.0.0
scikit_learn==1.7.2
streamlit==1.51.0
torch==2.9.0
